backend/requirements.txt
asgiref==3.10.0
boto3==1.40.59
botocore==1.40.59
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.6.0
django-debug-toolbar==6.0.0
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.1.0
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.14.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
